executor/requirements.txt
aiohttp
redis[hiredis]
python-dotenv
